Patch from Oliver Hookins:
"""
I've made a first-attempt patch against 0.13 since I couldn't get 0.12
building (some weird automake shit). There is definitely one problem,
which is the generating of any rules for the nat table, which need to be
specially sequenced to be generated after the headers for the nat
section, but before the COMMIT of that section. I had a brief glance
over filter.c but without taking weeks to figure it all out I don't know
 an immediate solution that you'll probably think of easily.

You might see fit to add this as a separate backend (e.g.
iptables-restore) rather than overwrite the existing iptables backend.

Enjoy!
"""