Support globbing in include directives.
"""
The attached patch implements globbing for include directives in filtergen
rulesets, along with glob()s standard sorting functionality.  This allows
filtergen to be more safe (not read backup files and the like scattered
around a directory) and also to nicely order files within a directory so you
can know that 99log.fg is going to be at the end of all the rules.

- Matt
"""