build: Convert `input/filtergen` to Bazel rules.
A direct translation of the SConscript rules into Bazel using flex and bison
rules to emit source, and compile into a single `cc_library`.  The rules could
be broken up into one per source file in the future, but this was done to make
tracking the conversion obvious for now.