build: Add coverage support to the dejagnu test rule.
A lot of hair-tearing, and then eventually finding the right keyword to search
for.

Thanks again phst:
https://github.com/phst/rules_elisp/blob/f9664cf0e5c86653bffd1a02b7c0cccb4d27e979/elisp/defs.bzl#L214

Also thanks
https://github.com/c-mita/bazel_issues/blob/d217e2b7e7dc71088b7e02ddaa25a8f6b0f7d880/coverage_common/README.md?plain=1#L47

The magic is documented with the comment, copied from phst.  We need to provide
the hidden coverage collection binaries and test environment pointing to them.

We also must return the instrumented files provider listing the attribbutes
that contain its sources.

Now a bazel coverage ... generaets a report!